requirements.txt
streamlit>=1.28.0
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0
python-dotenv>=1.0.0
openai>=1.3.0